+++ src/Sulu/Bundle/MediaBundle/Controller/MediaController.php
@@ -113,8 +113,10 @@ class MediaController extends AbstractMediaController implements ClassResourceIn
      */
     public function cgetAction(Request $request)
     {
+        $ids = array_filter(explode(',', $request->get('ids')));
+
         $fieldDescriptors = $this->getFieldDescriptors($this->getLocale($request), false);
-        $listBuilder = $this->getListBuilder($request, $fieldDescriptors);
+        $listBuilder = $this->getListBuilder($request, $fieldDescriptors, $ids);
         $listResponse = $listBuilder->execute();
 
         for ($i = 0, $length = count($listResponse); $i < $length; ++$i) {
@@ -134,6 +136,15 @@ class MediaController extends AbstractMediaController implements ClassResourceIn
                 $listResponse[$i]['name'],
                 $listResponse[$i]['version']
             );
+        }
+
+        if (0 < count($ids)) {
+            $result = [];
+            foreach ($listResponse as $item) {
+                $result[array_search($item['id'], $ids)] = $item;
+            }
+            ksort($result);
+            $listResponse = array_values($result);
         }
 
         $list = new ListRepresentation(
@@ -157,7 +168,7 @@ class MediaController extends AbstractMediaController implements ClassResourceIn
      *
      * @return DoctrineListBuilder
      */
-    private function getListBuilder(Request $request, array $fieldDescriptors)
+    private function getListBuilder(Request $request, array $fieldDescriptors, $ids)
     {
         $restHelper = $this->get('sulu_core.doctrine_rest_helper');
         $factory = $this->get('sulu_core.doctrine_list_builder_factory');
@@ -184,7 +195,6 @@ class MediaController extends AbstractMediaController implements ClassResourceIn
 
         // If no limit is set in request and limit is set by ids
         $requestLimit = $request->get('limit');
-        $ids = array_filter(explode(',', $request->get('ids')));
         $idsCount = count($ids);
 
         if ($idsCount > 0) {